restart service loop after an exception is raised
